readapted dataset class

+++ models/layers.py
@@ -580,9 +580,9 @@ class MelSpectrogram(nn.Module):
                 win_length=window_size,
                 hop_length=window_stride,
                 pad=window_stride // 2,
-            ).to(eegs.device)
+            ).to(eegs.device).float()
             eegs = einops.rearrange(eegs, "b s c -> b c s")
-            spectrogram = mel_fn(eegs.double()).float()  # (b c m s)
+            spectrogram = mel_fn(eegs)  # (b c m s)
         spectrogram = einops.rearrange(spectrogram, "b c m s -> b s c m")
         return spectrogram
